General refactor and optmization
User now lazily evaluates the following/followers list by not polling
the GitHub API unless the lists are requested. The follower/following
lists are now instantiated User objects which reduces the number of API
calls. The graph object also holds a set of processed users to ensure
that users are not processed twice.